requirements.txt
streamlit
openai
python-dotenv
pandas
openai
pytrends>=4.9.2
fastapi
uvicorn
axios